+++ src/Manager/ReaderManagerInterface.php
@@ -1,7 +1,7 @@
 <?php
 
 /*
- * Copyright (c) 2021 Heimrich & Hannot GmbH
+ * Copyright (c) 2022 Heimrich & Hannot GmbH
  *
  * @license LGPL-3.0-or-later
  */
@@ -130,7 +130,7 @@ interface ReaderManagerInterface
     /**
      * Get current twig environment.
      *
-     * @return Environment|\Twig_Environment
+     * @return Environment
      */
     public function getTwig();